requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml==6.0.2
tenacity==9.0.0
httpx==0.27.2
structlog==24.4.0
rich==13.9.2

# Kafka client
confluent-kafka==2.6.0

# Spark + Delta (same versions as Project 1 for cross-project cluster compat)
pyspark==3.5.3
delta-spark==3.2.1
deltalake==0.20.2

# ML — consume Project 1's registered models
mlflow==2.17.0
scikit-learn==1.5.2
lightgbm==4.5.0
numpy==1.26.4
pandas==2.2.3

# LLM — same abstraction as Project 2 (OpenAI-compatible + Anthropic native)
openai==1.54.0
anthropic==0.39.0

# Tracing
langfuse==2.55.0

# Serving / UI
fastapi==0.115.0
uvicorn[standard]==0.31.0
streamlit==1.39.0
plotly==5.24.1
prometheus-client==0.21.0

# Dev
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
black==24.10.0
ruff==0.7.1
